Extract MELS-style comma separated list to shared utility
This allows other UI components to use the same formatting for a long list.
