+++ src/i18n/strings/en_EN.json
@@ -253,6 +253,9 @@
     "This homeserver has exceeded one of its resource limits.": "This homeserver has exceeded one of its resource limits.",
     "Please <a>contact your service administrator</a> to continue using the service.": "Please <a>contact your service administrator</a> to continue using the service.",
     "Unable to connect to Homeserver. Retrying...": "Unable to connect to Homeserver. Retrying...",
+    "%(items)s and %(count)s others|other": "%(items)s and %(count)s others",
+    "%(items)s and %(count)s others|one": "%(items)s and one other",
+    "%(items)s and %(lastItem)s": "%(items)s and %(lastItem)s",
     "Your browser does not support the required cryptography extensions": "Your browser does not support the required cryptography extensions",
     "Not a valid Riot keyfile": "Not a valid Riot keyfile",
     "Authentication check failed: incorrect password?": "Authentication check failed: incorrect password?",
@@ -1046,9 +1049,6 @@
     "%(severalUsers)schanged their avatar %(count)s times|one": "%(severalUsers)schanged their avatar",
     "%(oneUser)schanged their avatar %(count)s times|other": "%(oneUser)schanged their avatar %(count)s times",
     "%(oneUser)schanged their avatar %(count)s times|one": "%(oneUser)schanged their avatar",
-    "%(items)s and %(count)s others|other": "%(items)s and %(count)s others",
-    "%(items)s and %(count)s others|one": "%(items)s and one other",
-    "%(items)s and %(lastItem)s": "%(items)s and %(lastItem)s",
     "collapse": "collapse",
     "expand": "expand",
     "Power level": "Power level",
